requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psutil==5.9.6
geocoder==1.38.1
Pillow==10.0.1
opencv-python-headless==4.8.1.78
requests==2.31.0
gunicorn==21.2.0